shitload of changes
including restructuring the divs so they are positioned properly,
instead of with javascript. fixing that problem with overlap of first
div, so now line works right away. Starting the animation of the menu.
fixing the footer so that it is easier to read, fixing broken footer
facebook/twitter images, fixing other font and animation issues